+++ yolort/runtime/__init__.py
@@ -1,5 +1,5 @@
 # Copyright (c) 2021, yolort team. All Rights Reserved.
-from .ort_modeling import PredictorORT
+from .y_onnxruntime import PredictorORT
 from .yolo_tensorrt_model import YOLOTRTModule
 
 __all__ = ["PredictorORT", "YOLOTRTModule"]
